+++ src/main/java/net/oneandone/pommes/cli/Base.java
@@ -39,7 +39,7 @@ public abstract class Base {
 
     public void run() throws Exception {
         try (Database database = environment.home.loadDatabase()) {
-            environment.imports(database);
+            environment.implicitScan(database);
             run(database);
         }
     }